LPS-142864 Reorganize assignment-related files
