+++ modules/dxp/apps/portal-workflow/portal-workflow-kaleo-designer-web/src/main/resources/META-INF/resources/designer/js/definition-builder/diagram-builder/components/sidebar/sections/assignments/select-assignment/ResourceActions.js
@@ -9,26 +9,23 @@
  * distribution rights of the Software.
  */
 
-import ClayButton from '@clayui/button';
-import PropTypes from 'prop-types';
+import ClayForm, {ClayInput} from '@clayui/form';
 import React from 'react';
 
-import SidebarPanel from '../SidebarPanel';
+import SidebarPanel from '../../../SidebarPanel';
 
-const Assignments = ({setContentName}) => (
-	<SidebarPanel panelTitle={Liferay.Language.get('assignments')}>
-		<ClayButton
-			className="mr-3"
-			displayType="secondary"
-			onClick={() => setContentName('assignments')}
-		>
-			{Liferay.Language.get('new')}
-		</ClayButton>
+const ResourceActions = () => (
+	<SidebarPanel panelTitle={Liferay.Language.get('resource-actions')}>
+		<ClayForm.Group>
+			<label htmlFor="resource-actions">
+				{Liferay.Language.get('resource-actions')}
+
+				<span className="ml-1 mr-1 text-warning">*</span>
+			</label>
+
+			<ClayInput component="textarea" id="resource-actions" type="text" />
+		</ClayForm.Group>
 	</SidebarPanel>
 );
 
-Assignments.propTypes = {
-	setContentName: PropTypes.func.isRequired,
-};
-
-export default Assignments;
+export default ResourceActions;
