+++ modules/dxp/apps/portal-workflow/portal-workflow-kaleo-designer-web/src/main/resources/META-INF/resources/designer/js/definition-builder/diagram-builder/components/sidebar/sections/sectionComponents.js
@@ -9,10 +9,10 @@
  * distribution rights of the Software.
  */
 
-import Assignments from './Assignments';
 import EdgeInformation from './EdgeInformation';
 import NodeInformation from './NodeInformation';
-import SelectAssignment from './SelectAssignment';
+import Assignments from './assignments/Assignments';
+import SelectAssignment from './assignments/SelectAssignment';
 
 const sectionComponents = {
 	assignments: Assignments,
